VPAAMP-782 static Xcode analysis - potential null pointer dereference
Reason for Change: address handful of compilation warnings

+++ isobmff/isobmffbox.cpp
@@ -82,7 +82,7 @@ void WriteUint64(uint8_t *dst, uint64_t val)
 {
 	uint32_t msw = (uint32_t)(val>>32);
 	WRITE_U32(dst, msw); dst+=4;
-	WRITE_U32(dst, val);
+	WRITE_U32(dst, static_cast<uint32_t>(val) );
 }
 
 /**
@@ -1035,7 +1035,7 @@ void TrunBox::setFirstSampleDuration(uint64_t sampleDuration)
 	duration = sampleDuration;
 	if (nullptr != first_sample_duration_loc)
 	{
-		WRITE_U32(first_sample_duration_loc, sampleDuration);
+		WRITE_U32(first_sample_duration_loc, static_cast<uint32_t>(sampleDuration) );
 	}
 }
 
@@ -1234,7 +1234,7 @@ void TfhdBox::setDefaultSampleDuration(uint64_t default_duration)
 	mDefaultSampleDuration = default_duration;
 	if (nullptr != default_sample_duration_location)
 	{
-		WRITE_U32(default_sample_duration_location, default_duration);
+		WRITE_U32(default_sample_duration_location, static_cast<uint32_t>(default_duration) );
 	}
 }
